Move Benchmarking classes to new project, 'RoughBench'
