+++ csharp/BSOA/RoughBench/FilesBenchmarker.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 
-namespace BSOA.Benchmarks.Diagnostics
+namespace RoughBench
 {
     public static class FilesBenchmarker
     {
